--- .frsh/Menu-KFEYW5NJ.js
@@ -1,1 +1,0 @@
-import{a as t}from"./chunk-LHEXP62T.js";import"./chunk-UXYBTFIZ.js";import"./chunk-U5OBHLK4.js";import{f as e}from"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import{c as i}from"./chunk-GGCEL3NW.js";function a({item:s}){return e("div",{class:"collapse collapse-plus",children:[e("input",{type:"checkbox"}),e("div",{class:"collapse-title",children:s.label}),e("div",{class:"collapse-content",children:e("ul",{children:[e("li",{children:e("a",{class:"underline text-sm",href:s.href,children:"Ver todos"})}),s.children?.map(l=>e("li",{children:e(a,{item:l})}))]})})]})}function p({items:s}){return e(i,{children:[e("ul",{class:"px-4 flex-grow flex flex-col divide-y divide-base-200",children:s.map(l=>e("li",{children:e(a,{item:l})}))}),e("ul",{class:"flex flex-col py-2 bg-base-200",children:[e("li",{children:e("a",{class:"flex items-center gap-4 px-4 py-2",href:"/wishlist",children:[e(t,{id:"Heart",width:20,height:20,strokeWidth:2}),e("span",{class:"text-sm",children:"Lista de desejos"})]})}),e("li",{children:e("a",{class:"flex items-center gap-4 px-4 py-2",href:"https://www.deco.cx",children:[e(t,{id:"MapPin",width:20,height:20,strokeWidth:2}),e("span",{class:"text-sm",children:"Nossas lojas"})]})}),e("li",{children:e("a",{class:"flex items-center gap-4 px-4 py-2",href:"https://www.deco.cx",children:[e(t,{id:"Phone",width:20,height:20,strokeWidth:2}),e("span",{class:"text-sm",children:"Fale conosco"})]})}),e("li",{children:e("a",{class:"flex items-center gap-4 px-4 py-2",href:"https://www.deco.cx",children:[e(t,{id:"User",width:20,height:20,strokeWidth:2}),e("span",{class:"text-sm",children:"Minha conta"})]})})]})]})}var o=p;export{o as default};
