--- .frsh/Searchbar-CHENNAG4.js
@@ -1,2 +1,0 @@
-import{a as F}from"./chunk-FHXDGFFW.js";import{a as z}from"./chunk-B72IEX5X.js";import{a as x}from"./chunk-MWXQMU46.js";import{a as $,b as R}from"./chunk-5MOTKOP3.js";import{b as A}from"./chunk-74TMIHW5.js";import{a as C}from"./chunk-E65QVBAJ.js";import{a as b}from"./chunk-CQCRCOLX.js";import{a as O}from"./chunk-4QO3HHI3.js";import"./chunk-VE4HZINV.js";import{a as w}from"./chunk-PBDGP2TB.js";import{a as y}from"./chunk-EYIWULCN.js";import{a as f}from"./chunk-LHEXP62T.js";import{a as h}from"./chunk-X2U7WVKR.js";import"./chunk-UXYBTFIZ.js";import"./chunk-U5OBHLK4.js";import"./chunk-U37GBM4D.js";import{a as v}from"./chunk-DSHYIDCS.js";import{a as E,b as k,f as e}from"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import{c as D}from"./chunk-GGCEL3NW.js";var H=(r,s)=>s.priceComponentType!=="https://schema.org/Installment"?r:!r||r.price>s.price?s:r.price<s.price?r:r.billingDuration&&s.billingDuration&&r.billingDuration<s.billingDuration?s:r,L=(r,s)=>{let{billingDuration:a,billingIncrement:i,price:t}=r;if(!a||!i)return"";let o=s<t;return`${a}x de R$ ${i} ${o?"com juros":"sem juros"}`},V=r=>{let s=r?.offers[0],a=s?.priceSpecification.find(n=>n.priceType==="https://schema.org/ListPrice"),i=s?.priceSpecification.reduce(H,null),t=s?.seller,o=s?.price,l=s?.availability;return{price:o,listPrice:a?.price,availability:l,seller:t,installments:i&&o?L(i,o):null}};var W=({url:r,isVariantOf:s})=>(s?.hasVariant??[]).flatMap(({additionalProperty:t=[],url:o})=>t.map(l=>({property:l,url:o}))).filter(t=>t.url).filter(t=>t.property.valueReference==="SPECIFICATION").sort((t,o)=>t.url<o.url?-1:t.url===o.url?0:1).reduce((t,{property:o,url:l})=>{let{name:n="",value:u=""}=o;return t[n]||(t[n]={}),t[n][u]||(t[n][u]=[]),l&&(l===r?t[n][u].unshift(l):t[n][u].push(l)),t},{});var X=r=>{let s=new URL(r);return`${s.pathname}${s.search}`},P=200,I=279;function J({product:r,preload:s,itemListName:a}){let{url:i,productID:t,name:o,image:l,offers:n,isVariantOf:u}=r,g=u?.productGroupID,[d,m]=l??[],{listPrice:c,price:p}=V(n),_=W(r),j=Object.entries(Object.values(_)[0]??{}),q={name:"select_item",params:{item_list_name:a,items:[A({product:r,price:p,listPrice:c})]}};return e("div",{class:"card card-compact card-bordered border-transparent hover:border-base-200 group w-full","data-deco":"view-product",id:`product-card-${t}`,...R(q),children:[e("figure",{class:"relative ",style:{aspectRatio:`${P} / ${I}`},children:[e("div",{class:"absolute top-0 right-0 z-10",children:e(z,{productGroupID:g,productID:t})}),e("a",{href:i&&X(i),"aria-label":"view product",class:"contents",children:[e(y,{src:d.url,alt:d.alternateName,width:P,height:I,class:"absolute transition-opacity rounded w-full opacity-100 group-hover:opacity-0",sizes:"(max-width: 640px) 50vw, 20vw",preload:s,loading:s?"eager":"lazy",decoding:"async"}),e(y,{src:m?.url??d.url,alt:m?.alternateName??d.alternateName,width:P,height:I,class:"absolute transition-opacity rounded w-full opacity-0 group-hover:opacity-100",sizes:"(max-width: 640px) 50vw, 20vw",loading:"lazy",decoding:"async"})]}),e("figcaption",{class:"glass card-body card-actions absolute bottom-0 left-0 w-full transition-opacity opacity-0 group-hover:opacity-100",children:e("ul",{class:"flex justify-center items-center gap-2 w-full",children:j.map(([M,[T]])=>e("a",{href:T,children:e(F,{variant:T===i?"active":"default",content:M})}))})})]}),e("div",{class:"card-body",children:[e("h2",{class:"card-title whitespace-nowrap overflow-hidden",children:o}),e("div",{class:"flex items-end gap-2",children:[e("span",{class:"line-through text-base-300 text-xs",children:x(c,n.priceCurrency)}),e("span",{class:"text-secondary",children:x(p,n.priceCurrency)})]})]})]})}var B=J;function G(r,s){let a=null,i=null,t=(...o)=>{t.clear(),i=()=>{t.clear(),r.call(t,...o)},a=setTimeout(i,s)};return t.clear=()=>{typeof a=="number"&&(clearTimeout(a),a=null,i=null)},t.flush=()=>{i?.()},Object.defineProperty(t,"pending",{get:()=>typeof a=="number"}),t}var N=v(null),S=v(!1),K=C.create("deco-sites/std/loaders/vtex/intelligentSearch/suggestions.ts"),Q=G(async r=>{try{N.value=await K({query:r,count:4})}catch(s){console.error(`Something went wrong while fetching suggestions 
-`,s)}finally{S.value=!1}},250),Y={setSearch:r=>{S.value=!0,Q(r)},loading:S,suggestions:N},U=()=>Y;function Z(){let{displaySearchbar:r}=O();return e(h,{class:"btn-ghost btn-circle",onClick:()=>r.value=!1,children:e(f,{id:"XMark",width:20,height:20,strokeWidth:2})})}function ee({placeholder:r="What are you looking for?",action:s="/s",name:a="q",query:i,variant:t="mobile"}){let o=k(null),{setSearch:l,suggestions:n,loading:u}=U(),g=!!n.value?.products?.length,d=!!n.value?.searches?.length,m=!g&&!d;return E(()=>{o.current&&o.current.focus()},[]),e("div",{class:"flex flex-col p-4 md:py-6 md:px-20",children:[e("div",{class:"flex items-center gap-4",children:[e("form",{id:"searchbar",action:s,class:"flex-grow flex gap-3 px-3 py-2 border border-base-200",children:[e(h,{class:"btn-ghost","aria-label":"Search",htmlFor:"searchbar",tabIndex:-1,children:e(f,{class:"text-base-300",id:"MagnifyingGlass",size:20,strokeWidth:.01})}),e("input",{ref:o,id:"search-input",class:"flex-grow outline-none placeholder-shown:sibling:hidden",name:a,defaultValue:i,onInput:c=>{let p=c.currentTarget.value;p&&$({name:"search",params:{search_term:p}}),l(p)},placeholder:r,role:"combobox","aria-controls":"search-suggestion",autocomplete:"off"}),e("button",{type:"button","aria-label":"Clean search",class:"focus:outline-none",tabIndex:-1,onClick:c=>{c.stopPropagation(),o.current!==null&&(o.current.value="",l(""))},children:e("span",{class:"text-sm",children:"limpar"})})]}),t==="desktop"&&e(Z,{})]}),e("div",{class:"flex flex-col gap-6 divide-y divide-base-200 mt-6 empty:mt-0 md:flex-row md:divide-y-0",children:m?e("div",{class:"py-16 md:py-6! flex flex-col gap-4 w-full",children:[e("span",{class:"font-medium text-xl text-center",role:"heading","aria-level":3,children:"Nenhum resultado encontrado"}),e("span",{class:"text-center text-base-300",children:"Vamos tentar de outro jeito? Verifique a ortografia ou use um termo diferente"})]}):e(D,{children:[e("div",{class:"flex flex-col gap-6 md:w-[15.25rem] md:max-w-[15.25rem]\\",children:[e("div",{class:"flex gap-2 items-center",children:[e("span",{class:"font-medium text-xl",role:"heading","aria-level":3,children:"Sugest\xF5es"}),u.value&&e(b,{})]}),e("ul",{id:"search-suggestion",class:"flex flex-col gap-6",children:n.value.searches?.map(({term:c})=>e("li",{children:e("a",{href:`/s?q=${c}`,class:"flex gap-4 items-center",children:[e("span",{children:e(f,{id:"MagnifyingGlass",size:20,strokeWidth:.01})}),e("span",{children:c})]})}))})]}),e("div",{class:"flex flex-col pt-6 md:pt-0 gap-6 overflow-x-hidden",children:[e("div",{class:"flex gap-2 items-center",children:[e("span",{class:"font-medium text-xl",role:"heading","aria-level":3,children:"Produtos sugeridos"}),u.value&&e(b,{})]}),e(w,{class:"carousel",children:n.value.products?.map((c,p)=>e(w.Item,{index:p,class:"carousel-item first:ml-4 last:mr-4 min-w-[200px] max-w-[200px]",children:e(B,{product:c})}))})]})]})})]})}var Ee=ee;export{Ee as default};
