--- .frsh/chunk-5MOTKOP3.js
@@ -1,1 +1,0 @@
-import{a as n}from"./chunk-U5OBHLK4.js";var e=t=>window.DECO_SITES_STD.sendAnalyticsEvent(t),o=t=>({onclick:n?()=>e(t):`(${e})(${JSON.stringify(t)})`});export{e as a,o as b};
