+++ .frsh/chunk-JC3BWQRB.js
@@ -1,0 +1,1 @@
+var c="__frsh_c",g=typeof document<"u";function i(s){if(!s.startsWith("/")||s.startsWith("//"))return s;try{let t=new URL(s,"https://freshassetcache.local");return t.protocol!=="https:"||t.host!=="freshassetcache.local"||t.searchParams.has(c)?s:(t.searchParams.set(c,"9e3286fb5fb5a50544e280095a28be1cbc8e03fe"),t.pathname+t.search+t.hash)}catch(t){return console.warn(`Failed to create asset() URL, falling back to regular path ('${s}'):`,t),s}}function h(s){if(s.includes("("))return s;let t=s.split(","),o=[];for(let n of t){let r=n.trimStart(),a=n.length-r.length;if(r==="")return s;let e=r.indexOf(" ");e===-1&&(e=r.length);let f=n.substring(0,a),l=r.substring(0,e),u=r.substring(e);o.push(f+i(l)+u)}return o.join(",")}function p(s){if(s.type==="img"||s.type==="source"){let{props:t}=s;if(t["data-fresh-disable-lock"])return;typeof t.src=="string"&&(t.src=i(t.src)),typeof t.srcset=="string"&&(t.srcset=h(t.srcset))}}export{g as a,i as b,p as c};
