--- .frsh/chunk-LHEXP62T.js
@@ -1,1 +1,0 @@
-import{b as t}from"./chunk-U5OBHLK4.js";import{f as r}from"./chunk-IAMGS6YL.js";function p({id:o,strokeWidth:s=16,size:e,width:i,height:n,...a}){return r("svg",{...a,width:i??e,height:n??e,strokeWidth:s,children:r("use",{href:t(`/sprites.svg#${o}`)})})}var c=p;export{c as a};
