+++ .frsh/chunk-QBUNRMND.js
@@ -1,0 +1,1 @@
+import{a as n}from"./chunk-JC3BWQRB.js";var e=t=>window.DECO_SITES_STD.sendAnalyticsEvent(t),o=t=>({onclick:n?()=>e(t):`(${e})(${JSON.stringify(t)})`});export{e as a,o as b};
