+++ .frsh/island-headersearchbar.js
@@ -1,1 +1,1 @@
-import{a as s}from"./chunk-TOUNTMGJ.js";import"./chunk-CQCRCOLX.js";import{a as t}from"./chunk-4QO3HHI3.js";import{b as o,c as a}from"./chunk-U37GBM4D.js";import"./chunk-DSHYIDCS.js";import{f as r}from"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import"./chunk-GGCEL3NW.js";var p="87px";var n=a(()=>import("./Searchbar-CHENNAG4.js"));function h({searchbar:i}){let{displaySearchbar:c}=t(),e=c.value&&window?.matchMedia?.("(min-width: 768px)")?.matches;return r("div",{class:`${e?"block border-y border-base-200 shadow":"hidden"} absolute left-0 top-0 w-screen z-50 bg-base-100`,style:{marginTop:p},children:e&&r(o,{fallback:r(s,{}),children:r(n,{...i,variant:"desktop"})})})}var d=h;export{d as default};
+import{a as s}from"./chunk-TOUNTMGJ.js";import"./chunk-CQCRCOLX.js";import{a as t}from"./chunk-4QO3HHI3.js";import{b as o,c as a}from"./chunk-U37GBM4D.js";import"./chunk-DSHYIDCS.js";import{f as r}from"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import"./chunk-GGCEL3NW.js";var p="87px";var n=a(()=>import("./Searchbar-W2QUKC3S.js"));function h({searchbar:i}){let{displaySearchbar:c}=t(),e=c.value&&window?.matchMedia?.("(min-width: 768px)")?.matches;return r("div",{class:`${e?"block border-y border-base-200 shadow":"hidden"} absolute left-0 top-0 w-screen z-50 bg-base-100`,style:{marginTop:p},children:e&&r(o,{fallback:r(s,{}),children:r(n,{...i,variant:"desktop"})})})}var d=h;export{d as default};
