+++ .frsh/island-productimagezoom.js
@@ -1,1 +1,1 @@
-import{a as e}from"./chunk-PBDGP2TB.js";import{a as u}from"./chunk-ZQJY2FQ5.js";import{a as c}from"./chunk-EYIWULCN.js";import{a as d}from"./chunk-HYV53KRH.js";import{a as o}from"./chunk-LHEXP62T.js";import{a}from"./chunk-X2U7WVKR.js";import"./chunk-UXYBTFIZ.js";import"./chunk-U5OBHLK4.js";import"./chunk-U37GBM4D.js";import{b as m}from"./chunk-DSHYIDCS.js";import{f as t}from"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import{c as n}from"./chunk-GGCEL3NW.js";var f="product-zoom";function b({images:p,width:i,height:l}){let r=m(!1);return t(n,{children:[t(a,{class:"hidden sm:inline-flex cursor-zoom-in btn-ghost",onClick:()=>r.value=!0,children:t(o,{id:"ArrowsPointingOut",size:20,stroke:"currentColor",strokeWidth:2})}),t(d,{loading:"lazy",mode:"center",open:r.value,onClose:()=>{r.value=!1},children:[t("div",{class:"relative",id:f,children:[t(e,{class:"carousel w-screen overflow-y-auto",children:p.map((s,h)=>t(e.Item,{index:h,class:"carousel-item w-full",children:t(c,{style:{aspectRatio:`${i} / ${l}`},src:s.url,alt:s.alternateName,width:i,height:l})}))}),t(e.PrevButton,{class:"btn btn-circle btn-outline absolute left-8 top-[50vh]",children:t(o,{size:20,id:"ChevronLeft",strokeWidth:3})}),t(e.NextButton,{class:"btn btn-circle btn-outline absolute right-8 top-[50vh]",children:t(o,{size:20,id:"ChevronRight",strokeWidth:3})})]}),t(u,{rootId:f})]})]})}var v=b;export{v as default};
+import{a as e}from"./chunk-PBDGP2TB.js";import{a as u}from"./chunk-ZQJY2FQ5.js";import{a as c}from"./chunk-EYIWULCN.js";import{a as d}from"./chunk-ZQF3XTRU.js";import{a as o}from"./chunk-FHJRSYQF.js";import{a}from"./chunk-X2U7WVKR.js";import"./chunk-UXYBTFIZ.js";import"./chunk-JC3BWQRB.js";import"./chunk-U37GBM4D.js";import{b as m}from"./chunk-DSHYIDCS.js";import{f as t}from"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import{c as n}from"./chunk-GGCEL3NW.js";var f="product-zoom";function b({images:p,width:i,height:l}){let r=m(!1);return t(n,{children:[t(a,{class:"hidden sm:inline-flex cursor-zoom-in btn-ghost",onClick:()=>r.value=!0,children:t(o,{id:"ArrowsPointingOut",size:20,stroke:"currentColor",strokeWidth:2})}),t(d,{loading:"lazy",mode:"center",open:r.value,onClose:()=>{r.value=!1},children:[t("div",{class:"relative",id:f,children:[t(e,{class:"carousel w-screen overflow-y-auto",children:p.map((s,h)=>t(e.Item,{index:h,class:"carousel-item w-full",children:t(c,{style:{aspectRatio:`${i} / ${l}`},src:s.url,alt:s.alternateName,width:i,height:l})}))}),t(e.PrevButton,{class:"btn btn-circle btn-outline absolute left-8 top-[50vh]",children:t(o,{size:20,id:"ChevronLeft",strokeWidth:3})}),t(e.NextButton,{class:"btn btn-circle btn-outline absolute right-8 top-[50vh]",children:t(o,{size:20,id:"ChevronRight",strokeWidth:3})})]}),t(u,{rootId:f})]})]})}var v=b;export{v as default};
