+++ .frsh/island-wishlistbutton.js
@@ -1,1 +1,1 @@
-import{a}from"./chunk-B72IEX5X.js";import"./chunk-E65QVBAJ.js";import"./chunk-VE4HZINV.js";import"./chunk-LHEXP62T.js";import"./chunk-X2U7WVKR.js";import"./chunk-UXYBTFIZ.js";import"./chunk-U5OBHLK4.js";import"./chunk-U37GBM4D.js";import"./chunk-DSHYIDCS.js";import"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import"./chunk-GGCEL3NW.js";export{a as default};
+import{a}from"./chunk-UOXEXUFU.js";import"./chunk-E65QVBAJ.js";import"./chunk-VE4HZINV.js";import"./chunk-FHJRSYQF.js";import"./chunk-X2U7WVKR.js";import"./chunk-UXYBTFIZ.js";import"./chunk-JC3BWQRB.js";import"./chunk-U37GBM4D.js";import"./chunk-DSHYIDCS.js";import"./chunk-IAMGS6YL.js";import"./chunk-CRNLDV6W.js";import"./chunk-GGCEL3NW.js";export{a as default};
