+++ sections/Banner/FullBanner.tsx
